+++ src/include/fsal.h
@@ -88,7 +88,7 @@ extern uint32_t root_op_export_set;
  * @brief node id used to construct recovery directory in
  * cluster implementation.
  */
-extern ushort g_nodeid;
+extern int g_nodeid;
 
 /**
  * @brief Ops context for asynch and not protocol tasks that need to use
